backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
redis==6.4.0
celery==5.5.3
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
